+++ lib/pushi/__init__.py
@@ -31,6 +31,7 @@ from . import base
 from . import event
 from . import subscription
 from . import web
+from . import web_push
 
 from .apn import APNAPI
 from .app import AppAPI
@@ -38,3 +39,4 @@ from .base import BASE_URL, BASE_WS_URL, API
 from .event import EventAPI
 from .subscription import SubscriptionAPI
 from .web import WebAPI
+from .web_push import WebPushAPI